misc: better packaging of css files
BREAKING CHANGE: Requires changing build scripts to include css-loader when using indivitual components
